Do not save all outputs to ctx.partial_evaluations, and instead recompute the last hierarchy level for all prefixes.
The reason here is that we expect the outputs to be sparse, i.e., most outputs of one hierarchy level will *not* be used as prefixes for the next hierarchy level. While this means computing some PRG evaluations twice, it saves a lot of time otherwise spent on protobuf serialization and deserialization.

Also consistently use absl::btree_map wherever we expect the inputs to be sorted.

PiperOrigin-RevId: 368040452